requirements.txt
google-genai
streamlit
pandas
numpy
openai
requests
pillow
python-dotenv
tqdm
langchain-google-genai
